+++ src/main.rs
@@ -224,6 +224,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     // 3D Tiles
     info!("Generating 3D Tiles tileset");
     let tileset_path = cli.output.join("tileset.json");
+    let tileset_path_pruned = cli.output.join("tileset_pruned.json");
     let mut tileset = formats::cesium3dtiles::Tileset::from_quadtree(
         &quadtree,
         &world,
@@ -247,7 +248,28 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
             // Tileset.make_implicit() outputs the tiles that have content. If only the leaves have
             //  content, then only the leaves are outputted.
             let tiles_subtrees =
-                tileset_implicit.make_implicit(&world.grid, &quadtree, cli.grid_export);
+                tileset_implicit.make_implicit(&world.grid, &quadtree, cli.grid_export, None);
+
+            info!("Writing 3D Tiles tileset");
+            tileset_implicit.to_file(&tileset_path)?;
+            info!("Writing subtrees for implicit tiling");
+            let subtrees_path = cli.output.join("subtrees");
+            fs::create_dir_all(&subtrees_path)?;
+            for (subtree_id, subtree_bytes) in &tiles_subtrees.1 {
+                fs::create_dir_all(
+                    subtrees_path.join(format!("{}/{}", subtree_id.level, subtree_id.x)),
+                )
+                .unwrap();
+                let out_path = subtrees_path
+                    .join(&subtree_id.to_string())
+                    .with_extension("subtree");
+                let mut subtree_file = File::create(&out_path)
+                    .unwrap_or_else(|_| panic!("could not create {:?} for writing", &out_path));
+                if let Err(e) = subtree_file.write_all(&subtree_bytes) {
+                    warn!("Failed to write subtree {} content", subtree_id);
+                }
+            }
+
             tiles_subtrees
         }
         false => {
@@ -259,6 +281,10 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
                 .into_iter()
                 .map(|tile_ref| (tile_ref.clone(), tile_ref.id.clone()))
                 .collect();
+
+            info!("Writing 3D Tiles tileset");
+            tileset.to_file(&tileset_path)?;
+
             (tiles, vec![])
         }
     };
@@ -693,9 +719,14 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
             //  because it is simpler than flipping the bits of the unavailable tiles,
             //  because of the mixed up explicit/implicit tile IDs. But ideally, we
             //  flip the bits, so we won't need to duplicate the tileset here.
-            let (_, subtrees) = tileset.make_implicit(&world.grid, &quadtree, cli.grid_export);
-            info!("Writing subtrees for implicit tiling");
-            let subtrees_path = cli.output.join("subtrees");
+            let (_, subtrees) = tileset.make_implicit(
+                &world.grid,
+                &quadtree,
+                cli.grid_export,
+                Some("subtrees_pruned"),
+            );
+            info!("Writing pruned subtrees for implicit tiling");
+            let subtrees_path = cli.output.join("subtrees_pruned");
             fs::create_dir_all(&subtrees_path)?;
             for (subtree_id, subtree_bytes) in subtrees {
                 fs::create_dir_all(
@@ -712,10 +743,9 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
                 }
             }
         }
+        info!("Writing pruned 3D Tiles tileset");
+        tileset.to_file(&tileset_path_pruned)?;
     }
-
-    info!("Writing 3D Tiles tileset");
-    tileset.to_file(&tileset_path)?;
 
     Ok(())
 }